split old GL utils file into separate pieces
