--- src/gl/gl.js
@@ -1,176 +1,0 @@
-// WebGL management and rendering functions
-import libtess from 'libtess';
-import log from 'loglevel';
-export var GL = {};
-
-// Setup a WebGL context
-// If no canvas element is provided, one is created and added to the document body
-GL.getContext = function getContext (canvas, options)
-{
-
-    var fullscreen = false;
-    if (canvas == null) {
-        canvas = document.createElement('canvas');
-        canvas.style.position = 'absolute';
-        canvas.style.top = 0;
-        canvas.style.left = 0;
-        canvas.style.zIndex = -1;
-        document.body.appendChild(canvas);
-        fullscreen = true;
-    }
-
-    var gl = canvas.getContext('experimental-webgl', options);
-    if (!gl) {
-        alert("Couldn't create WebGL context. Your browser probably doesn't support WebGL or it's turned off?");
-        throw "Couldn't create WebGL context";
-    }
-
-    GL.resizeCanvas(gl, window.innerWidth, window.innerHeight);
-    if (fullscreen === true) {
-        window.addEventListener('resize', function () {
-            GL.resizeCanvas(gl, window.innerWidth, window.innerHeight);
-        });
-    }
-
-    // VertexArrayObject.init(gl); // TODO: this pattern doesn't support multiple active GL contexts, should that even be supported?
-
-    return gl;
-};
-
-GL.resizeCanvas = function (gl, width, height)
-{
-    var device_pixel_ratio = window.devicePixelRatio || 1;
-    gl.canvas.style.width = width + 'px';
-    gl.canvas.style.height = height + 'px';
-    gl.canvas.width = Math.round(gl.canvas.style.width * device_pixel_ratio);
-    gl.canvas.height = Math.round(gl.canvas.style.width * device_pixel_ratio);
-    gl.viewport(0, 0, gl.canvas.width, gl.canvas.height);
-};
-
-// Compile & link a WebGL program from provided vertex and fragment shader sources
-// update a program if one is passed in. Create one if not. Alert and don't update anything if the shaders don't compile.
-GL.updateProgram = function GLupdateProgram (gl, program, vertex_shader_source, fragment_shader_source)
-{
-    try {
-        var vertex_shader = GL.createShader(gl, vertex_shader_source, gl.VERTEX_SHADER);
-        var fragment_shader = GL.createShader(gl, '#ifdef GL_ES\nprecision highp float;\n#endif\n\n' + fragment_shader_source, gl.FRAGMENT_SHADER);
-    }
-    catch(err) {
-        log.error(err);
-        throw err;
-    }
-
-    gl.useProgram(null);
-    if (program != null) {
-        var old_shaders = gl.getAttachedShaders(program);
-        for(var i = 0; i < old_shaders.length; i++) {
-            gl.detachShader(program, old_shaders[i]);
-        }
-    } else {
-        program = gl.createProgram();
-    }
-
-    if (vertex_shader == null || fragment_shader == null) {
-        return program;
-    }
-
-    gl.attachShader(program, vertex_shader);
-    gl.attachShader(program, fragment_shader);
-
-    gl.deleteShader(vertex_shader);
-    gl.deleteShader(fragment_shader);
-
-    gl.linkProgram(program);
-
-    if (!gl.getProgramParameter(program, gl.LINK_STATUS)) {
-        var program_error = new Error(
-            `WebGL program error:
-            VALIDATE_STATUS: ${gl.getProgramParameter(program, gl.VALIDATE_STATUS)}
-            ERROR: ${gl.getError()}
-            --- Vertex Shader ---
-            ${vertex_shader_source}
-            --- Fragment Shader ---
-            ${fragment_shader_source}`);
-        log.error(program_error);
-        throw program_error;
-    }
-
-    return program;
-};
-
-// Compile a vertex or fragment shader from provided source
-GL.createShader = function GLcreateShader (gl, source, type)
-{
-    var shader = gl.createShader(type);
-
-    gl.shaderSource(shader, source);
-    gl.compileShader(shader);
-
-    if (!gl.getShaderParameter(shader, gl.COMPILE_STATUS)) {
-        var shader_error =
-            "WebGL shader error:\n" +
-            (type === gl.VERTEX_SHADER ? "VERTEX" : "FRAGMENT") + " SHADER:\n" +
-            gl.getShaderInfoLog(shader);
-        throw shader_error;
-    }
-
-    return shader;
-};
-
-// Triangulation using libtess.js port of gluTesselator
-// https://github.com/brendankenny/libtess.js
-GL.tesselator = (function initTesselator() {
-    var tesselator = new libtess.GluTesselator();
-
-    // Called for each vertex of tesselator output
-    function vertexCallback(data, polyVertArray) {
-        polyVertArray.push([data[0], data[1]]);
-    }
-
-    // Called when segments intersect and must be split
-    function combineCallback(coords, data, weight) {
-        return coords;
-    }
-
-    // Called when a vertex starts or stops a boundary edge of a polygon
-    function edgeCallback(flag) {
-        // No-op callback to force simple triangle primitives (no triangle strips or fans).
-        // See: http://www.glprogramming.com/red/chapter11.html
-        // "Since edge flags make no sense in a triangle fan or triangle strip, if there is a callback
-        // associated with GLU_TESS_EDGE_FLAG that enables edge flags, the GLU_TESS_BEGIN callback is
-        // called only with GL_TRIANGLES."
-        log.trace('GL.tesselator: edge flag: ' + flag);
-    }
-
-    tesselator.gluTessCallback(libtess.gluEnum.GLU_TESS_VERTEX_DATA, vertexCallback);
-    tesselator.gluTessCallback(libtess.gluEnum.GLU_TESS_COMBINE, combineCallback);
-    tesselator.gluTessCallback(libtess.gluEnum.GLU_TESS_EDGE_FLAG, edgeCallback);
-
-    // Brendan Kenny:
-    // libtess will take 3d verts and flatten to a plane for tesselation
-    // since only doing 2d tesselation here, provide z=1 normal to skip
-    // iterating over verts only to get the same answer.
-    // comment out to test normal-generation code
-    tesselator.gluTessNormal(0, 0, 1);
-
-    return tesselator;
-})();
-
-GL.triangulatePolygon = function GLTriangulate (contours)
-{
-    var triangleVerts = [];
-    GL.tesselator.gluTessBeginPolygon(triangleVerts);
-
-    for (var i = 0; i < contours.length; i++) {
-        GL.tesselator.gluTessBeginContour();
-        var contour = contours[i];
-        for (var j = 0; j < contour.length; j ++) {
-            var coords = [contour[j][0], contour[j][1], 0];
-            GL.tesselator.gluTessVertex(coords, coords);
-        }
-        GL.tesselator.gluTessEndContour();
-    }
-
-    GL.tesselator.gluTessEndPolygon();
-    return triangleVerts;
-};
